remove old sdr scripts
